Adjust directory structure
